+++ packages/drawer/src/DrawerToolbarLayout/DrawerToolbarLayout.stories.tsx
@@ -17,7 +17,7 @@ import { Body } from '@leafygreen-ui/typography';
 import { DisplayMode } from '../Drawer';
 import { Size } from '../Drawer/Drawer.types';
 import { DrawerLayout, DrawerLayoutProps } from '../DrawerLayout';
-import { useDrawerToolbarContext } from '../DrawerToolbarContext';
+import { useDrawerToolbarContext } from './DrawerToolbarContext/DrawerToolbarContext';
 
 const SEED = 0;
 faker.seed(SEED);
