+++ packages/drawer/src/DrawerToolbarLayout/DrawerToolbarLayoutContainer.tsx
@@ -4,8 +4,8 @@ import { Toolbar, ToolbarIconButton } from '@leafygreen-ui/toolbar';
 
 import { Drawer } from '../Drawer/Drawer';
 import { useDrawerLayoutContext } from '../DrawerLayout';
-import { useDrawerToolbarContext } from '../DrawerToolbarContext';
-import { DrawerWithToolbarWrapper } from '../DrawerWithToolbarWrapper';
+import { useDrawerToolbarContext } from './DrawerToolbarContext/DrawerToolbarContext';
+import { DrawerWithToolbarWrapper } from './DrawerWithToolbarWrapper/DrawerWithToolbarWrapper';
 import { LayoutComponent } from '../LayoutComponent';
 import { DEFAULT_LGID_ROOT, getLgIds } from '../utils';
