+++ packages/drawer/src/DrawerToolbarLayout/DrawerWithToolbarWrapper/DrawerWithToolbarWrapper.tsx
@@ -2,7 +2,7 @@ import React, { forwardRef, useEffect, useState } from 'react';
 
 import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
 
-import { useDrawerLayoutContext } from '../DrawerLayout';
+import { useDrawerLayoutContext } from '../../DrawerLayout';
 
 import { getDrawerWithToolbarWrapperStyles } from './DrawerWithToolbarWrapper.styles';
 import { DrawerWithToolbarWrapperProps } from './DrawerWithToolbarWrapper.types';
